1、通过user_id查询电子通信方式、地址、照片、学生亲属表的数据

+++ src/main/java/edu/uni/userBaseInfo1/controller/AcademicController.java
@@ -62,9 +62,9 @@ public class AcademicController {
     //@ResponseBody：指明该方法效果等同于通过response对象输出指定格式的数据（JSON）
     @ApiOperation( value = "以一个id获取一条受教育信息记录详情",notes = "未测试" )
     @GetMapping("academic/{id}")
-    @ApiImplicitParam(name = "id", value = "Academic表的一个id", required = false, dataType = "Integer" , paramType = "path")
+    @ApiImplicitParam(name = "id", value = "Academic表的一个id", required = false, dataType = "Long" , paramType = "path")
     @ResponseBody
-    public void receive(@PathVariable Integer id, HttpServletResponse response) throws IOException {
+    public void receive(@PathVariable Long id, HttpServletResponse response) throws IOException {
         //设置返回的数据格式
         response.setContentType("application/json;charset=utf-8");
         //拼接缓存键名（字符串）
@@ -87,10 +87,6 @@ public class AcademicController {
 
     }
 
-    /**
-     * Author: laizhouhao 10:29 2019/4/30
-     * @apiNote: 查询所有学生信息记录
-     */
     /**
      * Author: laizhouhao 15:19 2019/4/30
      * @apiNote: 查询所有受教育信息记录
@@ -142,10 +138,10 @@ public class AcademicController {
      * @return 删除操作结果
      */
     @ApiOperation(value="删除受教育程度信息", notes="未测试")
-    @ApiImplicitParam(name = "id", value = "受教育程度id", required = true, dataType = "Integer", paramType = "path")
+    @ApiImplicitParam(name = "id", value = "受教育程度id", required = true, dataType = "Long", paramType = "path")
     @DeleteMapping("/academic/{id}")   //delete请求
     @ResponseBody
-    public Result destroy(@PathVariable Integer id){
+    public Result destroy(@PathVariable Long id){
         boolean success = academicService.delete(id);
         if(success){
             // 清空相关缓存
