+++ src/main/java/edu/uni/userBaseInfo1/controller/AddrAreaController.java
@@ -61,9 +61,9 @@ public class AddrAreaController {
     //@ResponseBody：指明该方法效果等同于通过response对象输出指定格式的数据（JSON）
     @ApiOperation( value = "以一个id获取一条城区记录详情",notes = "2019-5-2 11:00:34已通过测试" )
     @GetMapping("addrArea/{id}")
-    @ApiImplicitParam(name = "id", value = "AddrArea表的一个id", required = false, dataType = "Integer" , paramType = "path")
+    @ApiImplicitParam(name = "id", value = "AddrArea表的一个id", required = false, dataType = "Long" , paramType = "path")
     @ResponseBody
-    public void receive(@PathVariable Integer id, HttpServletResponse response) throws IOException {
+    public void receive(@PathVariable Long id, HttpServletResponse response) throws IOException {
         //设置返回的数据格式
         response.setContentType("application/json;charset=utf-8");
         //拼接缓存键名（字符串）
@@ -136,10 +136,10 @@ public class AddrAreaController {
      * @apiNote: 删除城区
      */
     @ApiOperation(value="删除城区", notes="2019-5-2 11:01:36 已通过测试")
-    @ApiImplicitParam(name = "id", value = "城区的id", required = true, dataType = "Integer", paramType = "path")
+    @ApiImplicitParam(name = "id", value = "城区的id", required = true, dataType = "Long", paramType = "path")
     @DeleteMapping("/addrArea/{id}")   //delete请求
     @ResponseBody
-    public Result destroy(@PathVariable Integer id){
+    public Result destroy(@PathVariable Long id){
         boolean success = addrAreaService.deleteAddrArea(id);
         if(success){
             // 清空相关缓存
