+++ src/main/java/edu/uni/userBaseInfo1/controller/AddrCountryController.java
@@ -60,9 +60,9 @@ public class AddrCountryController {
     //@ResponseBody：指明该方法效果等同于通过response对象输出指定格式的数据（JSON）
     @ApiOperation( value = "以一个id获取一条国家记录详情",notes = "2019-5-2 11:02:32 已通过测试" )
     @GetMapping("addrCountry/{id}")
-    @ApiImplicitParam(name = "id", value = "AddrCountry表的一个id", required = false, dataType = "Integer" , paramType = "path")
+    @ApiImplicitParam(name = "id", value = "AddrCountry表的一个id", required = false, dataType = "Long" , paramType = "path")
     @ResponseBody
-    public void receive(@PathVariable Integer id, HttpServletResponse response) throws IOException {
+    public void receive(@PathVariable Long id, HttpServletResponse response) throws IOException {
         //设置返回的数据格式
         response.setContentType("application/json;charset=utf-8");
         //拼接缓存键名（字符串）
@@ -135,10 +135,10 @@ public class AddrCountryController {
      * @apiNote: 删除国家
      */
     @ApiOperation(value="删除国家", notes="2019-5-2 11:02:50 已通过测试")
-    @ApiImplicitParam(name = "id", value = "国家的id", required = true, dataType = "Integer", paramType = "path")
+    @ApiImplicitParam(name = "id", value = "国家的id", required = true, dataType = "Long", paramType = "path")
     @DeleteMapping("/addrCountry/{id}")   //delete请求
     @ResponseBody
-    public Result destroy(@PathVariable Integer id){
+    public Result destroy(@PathVariable Long id){
         boolean success = addrCountryService.deleteAddrCountry(id);
         if(success){
             // 清空相关缓存
