+++ src/main/java/edu/uni/userBaseInfo1/controller/EmployeeController.java
@@ -60,9 +60,9 @@ public class EmployeeController {
     //@ResponseBody：指明该方法效果等同于通过response对象输出指定格式的数据（JSON）
     @ApiOperation( value = "以一个id获取一条职员记录详情",notes = "2019-5-2 11:05:35已通过测试" )
     @GetMapping("employee/{id}")
-    @ApiImplicitParam(name = "id", value = "Employee表的一个id", required = false, dataType = "Integer" , paramType = "path")
+    @ApiImplicitParam(name = "id", value = "Employee表的一个id", required = false, dataType = "Long" , paramType = "path")
     @ResponseBody
-    public void receive(@PathVariable Integer id, HttpServletResponse response) throws IOException {
+    public void receive(@PathVariable Long id, HttpServletResponse response) throws IOException {
         //设置返回的数据格式
         response.setContentType("application/json;charset=utf-8");
         //拼接缓存键名（字符串）
@@ -135,10 +135,10 @@ public class EmployeeController {
      * @apiNote: 删除职员
      */
     @ApiOperation(value="删除职员", notes="2019-5-2 11:05:51已通过测试")
-    @ApiImplicitParam(name = "id", value = "职员的id", required = true, dataType = "Integer", paramType = "path")
+    @ApiImplicitParam(name = "id", value = "职员的id", required = true, dataType = "Long", paramType = "path")
     @DeleteMapping("/employee/{id}")   //delete请求
     @ResponseBody
-    public Result destroy(@PathVariable Integer id){
+    public Result destroy(@PathVariable Long id){
         boolean success = employeeService.deleteEmployee(id);
         if(success){
             // 清空相关缓存
