+++ src/main/java/edu/uni/userBaseInfo1/controller/FirstLevelDisciplineController.java
@@ -61,9 +61,9 @@ public class FirstLevelDisciplineController {
     //@ResponseBody：指明该方法效果等同于通过response对象输出指定格式的数据（JSON）
     @ApiOperation( value = "以一个id获取一条一级学科记录详情",notes = "2019-5-2 11:06:51已通过测试" )
     @GetMapping("firstLevelDiscipline/{id}")
-    @ApiImplicitParam(name = "id", value = "firstLevelDiscipline表的一个id", required = false, dataType = "Integer" , paramType = "path")
+    @ApiImplicitParam(name = "id", value = "firstLevelDiscipline表的一个id", required = false, dataType = "Long" , paramType = "path")
     @ResponseBody
-    public void receive(@PathVariable Integer id, HttpServletResponse response) throws IOException {
+    public void receive(@PathVariable Long id, HttpServletResponse response) throws IOException {
         //设置返回的数据格式
         response.setContentType("application/json;charset=utf-8");
         //拼接缓存键名（字符串）
@@ -136,10 +136,10 @@ public class FirstLevelDisciplineController {
      * @apiNote: 删除一级学科
      */
     @ApiOperation(value="删除一级学科", notes="2019-5-2 11:06:37已通过测试")
-    @ApiImplicitParam(name = "id", value = "一级学科的id", required = true, dataType = "Integer", paramType = "path")
+    @ApiImplicitParam(name = "id", value = "一级学科的id", required = true, dataType = "Long", paramType = "path")
     @DeleteMapping("/firstLevelDiscipline/{id}")   //delete请求
     @ResponseBody
-    public Result destroy(@PathVariable Integer id){
+    public Result destroy(@PathVariable Long id){
         boolean success = firstLevelDisciplineService.deleteFirstLevelDiscipline(id);
         if(success){
             // 清空相关缓存
