+++ src/main/java/edu/uni/userBaseInfo1/controller/UserController.java
@@ -60,9 +60,9 @@ public class UserController {
     //@ResponseBody：指明该方法效果等同于通过response对象输出指定格式的数据（JSON）
     @ApiOperation( value = "以一个id获取一条用户记录详情",notes = "2019-5-2 11:08:41已通过测试" )
     @GetMapping("user/{id}")
-    @ApiImplicitParam(name = "id", value = "User表的一个id", required = false, dataType = "Integer" , paramType = "path")
+    @ApiImplicitParam(name = "id", value = "User表的一个id", required = false, dataType = "Long" , paramType = "path")
     @ResponseBody
-    public void receive(@PathVariable Integer id, HttpServletResponse response) throws IOException {
+    public void receive(@PathVariable Long id, HttpServletResponse response) throws IOException {
         //设置返回的数据格式
         response.setContentType("application/json;charset=utf-8");
         //拼接缓存键名（字符串）
@@ -135,10 +135,10 @@ public class UserController {
      * @apiNote: 删除用户
      */
     @ApiOperation(value="删除用户", notes="2019-5-2 11:08:55已通过测试")
-    @ApiImplicitParam(name = "id", value = "用户的id", required = true, dataType = "Integer", paramType = "path")
+    @ApiImplicitParam(name = "id", value = "用户的id", required = true, dataType = "Long", paramType = "path")
     @DeleteMapping("/user/{id}")   //delete请求
     @ResponseBody
-    public Result destroy(@PathVariable Integer id){
+    public Result destroy(@PathVariable Long id){
         boolean success = userService.deleteUser(id);
         if(success){
             // 清空相关缓存
