+++ src/main/java/edu/uni/userBaseInfo1/controller/UserUploadFileController.java
@@ -63,9 +63,9 @@ public class UserUploadFileController {
     //@ResponseBody：指明该方法效果等同于通过response对象输出指定格式的数据（JSON）
     @ApiOperation( value = "以一个id获取一条用户上传文件记录详情",notes = "2019-5-2 11:10:10已通过测试" )
     @GetMapping("userUploadFile/{id}")
-    @ApiImplicitParam(name = "id", value = "userUploadFile表的一个id", required = false, dataType = "Integer" , paramType = "path")
+    @ApiImplicitParam(name = "id", value = "userUploadFile表的一个id", required = false, dataType = "Long" , paramType = "path")
     @ResponseBody
-    public void receive(@PathVariable Integer id, HttpServletResponse response) throws IOException {
+    public void receive(@PathVariable Long id, HttpServletResponse response) throws IOException {
         //设置返回的数据格式
         response.setContentType("application/json;charset=utf-8");
         //拼接缓存键名（字符串）
@@ -138,10 +138,10 @@ public class UserUploadFileController {
      * @apiNote: 删除用户上传文件
      */
     @ApiOperation(value="删除用户上传文件", notes="2019-5-2 11:10:24已通过测试")
-    @ApiImplicitParam(name = "id", value = "用户上传文件的id", required = true, dataType = "Integer", paramType = "path")
+    @ApiImplicitParam(name = "id", value = "用户上传文件的id", required = true, dataType = "Long", paramType = "path")
     @DeleteMapping("/userUploadFile/{id}")   //delete请求
     @ResponseBody
-    public Result destroy(@PathVariable Integer id){
+    public Result destroy(@PathVariable Long id){
         boolean success = userUploadFileService.deleteUserUploadFile(id);
         if(success){
             // 清空相关缓存
