+++ src/main/java/edu/uni/userBaseInfo1/controller/UserinfoApplyController.java
@@ -61,9 +61,9 @@ public class UserinfoApplyController {
     //@ResponseBody：指明该方法效果等同于通过response对象输出指定格式的数据（JSON）
     @ApiOperation( value = "以一个id获取一条用户信息申请记录详情",notes = "2019-5-2 11:09:42已通过测试" )
     @GetMapping("userinfoApply/{id}")
-    @ApiImplicitParam(name = "id", value = "userinfoApply表的一个id", required = false, dataType = "Integer" , paramType = "path")
+    @ApiImplicitParam(name = "id", value = "userinfoApply表的一个id", required = false, dataType = "Long" , paramType = "path")
     @ResponseBody
-    public void receive(@PathVariable Integer id, HttpServletResponse response) throws IOException {
+    public void receive(@PathVariable Long id, HttpServletResponse response) throws IOException {
         //设置返回的数据格式
         response.setContentType("application/json;charset=utf-8");
         //拼接缓存键名（字符串）
@@ -136,10 +136,10 @@ public class UserinfoApplyController {
      * @apiNote: 删除用户信息申请
      */
     @ApiOperation(value="删除用户信息申请", notes="2019-5-2 11:09:56已通过测试")
-    @ApiImplicitParam(name = "id", value = "用户信息申请的id", required = true, dataType = "Integer", paramType = "path")
+    @ApiImplicitParam(name = "id", value = "用户信息申请的id", required = true, dataType = "Long", paramType = "path")
     @DeleteMapping("/userinfoApply/{id}")   //delete请求
     @ResponseBody
-    public Result destroy(@PathVariable Integer id){
+    public Result destroy(@PathVariable Long id){
         boolean success = userinfoApplyService.deleteUserinfoApply(id);
         if(success){
             // 清空相关缓存
